fix: 리포트 대기 화면 비활성화 이벤트 수정, 파일이름 변경
